Time: 74 ms (14.60%) | Memory: 56.5 MB (56.40%) - LeetSync

+++ 2581-divide-players-into-teams-of-equal-skill/divide-players-into-teams-of-equal-skill.cpp
@@ -17,7 +17,7 @@ public:
             if(ex!=(skill[i]+skill[j])) {
                 return -1;
             }
-            ans+=(long long)skill[i]*skill[j];
+            ans+=skill[i]*skill[j];
             i++;
             j--;
         }
